+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ CHANGELOG
 - Upgrade mypy
   [vangheem]
 
+- Fix not setting cache values for updated object when push is not enabled
+  [vangheem]
+
+- Fix conflict error handling with registry objects
+  [vangheem]
+
 - Sorted imports in all files and added `isort` in .travis to keep the format
   [masipcat]